+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres (more or less) to [Semantic Versioning](http://semver.o
 
 ## Unreleased
 * Added X and Y dragging on calender component
+* Restored diagonal scrolling with trackpad
+* fixed bug with shuftKey modifier when scrolling
 
 ## 0.28.0